committing work done on updating distributors profile

+++ lib/core/connect_end/model/get_incoming_order_ddetail_response_model/item.dart
@@ -13,7 +13,7 @@ class Item {
   String? sku;
   String? imageUrl;
   int? quantity;
-  int? unitPrice;
+  double? unitPrice;
   int? lineTotal;
   int? minimumOrderQuantity;
   int? packSize;
